stash implement ConstParam

+++ src/main.rs
@@ -78,7 +78,7 @@ fn update(_app: &App, _model: &mut Model, _update: Update) {
     egui.set_elapsed_time(_update.since_start);
 
     let ctx = egui.begin_frame();
-    crate::ui::style::apply_custom_style(&ctx, _model.global_settings.window_opacity.value);
+    crate::ui::style::apply_custom_style(&ctx, _model.global_settings.window_opacity.value as u8);
 
     // Modular Windows with this
     // egui::Window::new("Global Settings").show(&ctx, |ui| {
